LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/common-to-both-axes-supercedes-first-in-tree-order-expected.txt

Box 0Box 1Box 2Box 3Box 4Box 5Box 6Box 7Box 8Box 9

FAIL box0 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box0 in x axis after layout change expected 200 but got 100
FAIL box1 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box1 in x axis after layout change expected 200 but got 100
FAIL box2 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box2 in x axis after layout change expected 200 but got 100
FAIL box3 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box3 in x axis after layout change expected 200 but got 100
FAIL box4 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box4 in x axis after layout change expected 200 but got 100
FAIL box5 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box5 in x axis after layout change expected 200 but got 100
FAIL box6 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box6 in x axis after layout change expected 200 but got 100
FAIL box7 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box7 in x axis after layout change expected 200 but got 100
FAIL box8 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box8 in x axis after layout change expected 200 but got 100
FAIL box9 is common to both axes and is the snap target despite being last in tree order. assert_equals: scroller followed box9 in x axis after layout change expected 200 but got 100
